+++ lib/decidim/kids/engine.rb
@@ -44,14 +44,14 @@ module Decidim
         Decidim::System::UpdateOrganizationForm.include(Decidim::Kids::System::OrganizationFormOverride)
         Decidim::System::UpdateOrganization.include(Decidim::Kids::System::UpdateOrganizationOverride)
         Decidim::System::RegisterOrganization.include(Decidim::Kids::System::RegisterOrganizationOverride)
-        Decidim::ImpersonateUsers.include(Decidim::Kids::ImpersonateUsersOverride)
       end
 
-      # initializer "decidim_kids.overrides", after: "decidim.action_controller" do
-      #   config.to_prepare do
-      #     # Controller overrides here
-      #   end
-      # end
+      initializer "decidim_kids.overrides", after: "decidim.action_controller" do
+        config.to_prepare do
+          Decidim::ApplicationController.include(Decidim::Kids::ImpersonateMinors)
+          Decidim::DeviseControllers.include(Decidim::Kids::ImpersonateMinors)
+        end
+      end
 
       initializer "decidim_kids.webpacker.assets_path" do
         Decidim.register_assets_path File.expand_path("app/packs", root)
